pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "google-generativeai>=0.8.6",
    "litellm>=1.80.16",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
]

[tool.pytest.ini_options]
# pytest 配置
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
# Ruff 配置
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []
